% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac.R
\name{ac_count}
\alias{ac_count}
\title{Get the total record count for a WPRDC resource}
\usage{
ac_count(resource_id)
}
\arguments{
\item{resource_id}{CKAN resource ID}
}
\value{
integer record count
}
\description{
Get the total record count for a WPRDC resource
}
