% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac.R
\name{ac_metadata}
\alias{ac_metadata}
\title{Get metadata for a WPRDC dataset}
\usage{
ac_metadata(id_or_name)
}
\arguments{
\item{id_or_name}{Dataset ID or name slug}
}
\value{
tibble with dataset metadata including resource URLs
}
\description{
Get metadata for a WPRDC dataset
}
